fix input fields in dialog boxes sometimes not being selected properly
This issue occurs when repeatedly placing bookmarks. The input box doesn't get selected even though the method is called.
I'm assuming this happens because the same game object is selected twice in a row, resulting in the events not being sent properly.
A working solution seems to be to set the selected game object to null first.